third_party/blink/web_tests/svg/dynamic-updates/SVGFilterElement-dom-filterUnits-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'filterUnits' attribute of the SVGFilterElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS filterElement.getAttribute('filterUnits') is "userSpaceOnUse"
PASS filterElement.getAttribute('filterUnits') is "objectBoundingBox"
PASS successfullyParsed is true

TEST COMPLETE
